requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
pytest==8.3.3
httpx==0.27.2
